Refactor and tie in of signal events
